Engine: fixed goal inheritance handling for SimpleContinuations. Extended compile-time contract checking.
Previously SimpleContinuation did not inherite the goal from the creating query (QQ / QQQ), so the goal was lost and done processsing could not propagate back to it. QQQQ (previously known as ContinuationQuery) handles this now.

QQQQ_ has been introduced to handle cases where inheritance handling is not desired (i.e. for top-level SimpleContinuations that are not assiciated with any goal).

An unrelated change is that Continuation.done is now required to return ContinuationRequestorCalledToken too.